refactor(160): move puzzle assignment into assignment package
