+++ app/src/main/java/org/cescfe/numpairs/feature/game/presentation/GameUiStateFactory.kt
@@ -1,7 +1,9 @@
 package org.cescfe.numpairs.feature.game.presentation
 
 import org.cescfe.numpairs.domain.puzzle.LiveStripEntryOperatorUsage
-import org.cescfe.numpairs.domain.puzzle.StripEntryUsageByOperator
+import org.cescfe.numpairs.domain.puzzle.assignment.StripEntryUsageByOperator
+import org.cescfe.numpairs.domain.puzzle.assignment.operandSelectionChoicesFor
+import org.cescfe.numpairs.domain.puzzle.assignment.stripEntryUsageByOperator
 import org.cescfe.numpairs.domain.puzzle.liveRuleConflictsByStripEntryOperator
 import org.cescfe.numpairs.domain.puzzle.liveRuleConflictsByTile
 import org.cescfe.numpairs.domain.puzzle.mismatchedSumProductPairingTileIndexes
@@ -11,8 +13,6 @@ import org.cescfe.numpairs.domain.puzzle.model.PuzzleCompletionState
 import org.cescfe.numpairs.domain.puzzle.model.StripItem
 import org.cescfe.numpairs.domain.puzzle.model.Tile
 import org.cescfe.numpairs.domain.puzzle.model.TileResolutionState
-import org.cescfe.numpairs.domain.puzzle.operandSelectionChoicesFor
-import org.cescfe.numpairs.domain.puzzle.stripEntryUsageByOperator
 
 internal object GameUiStateFactory {
     private val availableOperators = listOf(
