+++ app/src/main/java/org/cescfe/numpairs/feature/game/presentation/GameViewModel.kt
@@ -5,13 +5,13 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import org.cescfe.numpairs.data.puzzle.seed.initialPuzzle as defaultInitialPuzzle
+import org.cescfe.numpairs.domain.puzzle.assignment.operandSelectionChoicesFor
 import org.cescfe.numpairs.domain.puzzle.model.Board
 import org.cescfe.numpairs.domain.puzzle.model.OperandSlot
 import org.cescfe.numpairs.domain.puzzle.model.Operator
 import org.cescfe.numpairs.domain.puzzle.model.Puzzle
 import org.cescfe.numpairs.domain.puzzle.model.StripItem
 import org.cescfe.numpairs.domain.puzzle.model.Tile
-import org.cescfe.numpairs.domain.puzzle.operandSelectionChoicesFor
 
 class GameViewModel(initialPuzzle: Puzzle = defaultInitialPuzzle) : ViewModel() {
     private var puzzle: Puzzle = initialPuzzle
