+++ app/src/test/java/org/cescfe/numpairs/domain/puzzle/PuzzleAssignmentsTest.kt
@@ -1,5 +1,8 @@
 package org.cescfe.numpairs.domain.puzzle
 
+import org.cescfe.numpairs.domain.puzzle.assignment.IndexedResolvedTileAssignment
+import org.cescfe.numpairs.domain.puzzle.assignment.ResolvedOperandAssignment
+import org.cescfe.numpairs.domain.puzzle.assignment.ResolvedTileAssignment
 import org.cescfe.numpairs.domain.puzzle.model.Operator
 import org.junit.Assert.assertThrows
 import org.junit.Test
